SM-1509 - Fake API backend for easy integration tests (#1262)
## 🎟️ Tracking

https://bitwarden.atlassian.net/browse/SM-1509

## 📔 Objective

Testing on an official Bitwarden server requires manual creation of a
user account, an organization (with additional licensing hurdles for
non-cloud/dev environments), projects, secrets, and access tokens.

This implements a minimal API server that only requires an access token
that we can safely share (because it is for a fake account that only
exists on this fake server) to quickly test basic CRUD functionality of
the Secrets Manager clients.

## 🖼️ Screenshots

See [Jira ticket](https://bitwarden.atlassian.net/browse/SM-1509) for
overview. Unfortunately, the file size was too large to upload it here
directly.

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes